pyproject.toml
[tool.poetry]
name = "turtamqtt"
version = "0.1.0"
description = ""
authors = ["Dogan ZORLU <doganzorlu@doganzorlu.com>"]

[tool.poetry.dependencies]
python = "^3.10"
turta-iothat = "^1.1.1"
smbus = "^1.1.post2"
paho-mqtt = "^1.6.1"
"RPi.GPIO" = "^0.7.1"
confuse = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
